Using util.inherits
