+++ 03 Events/01 Events/README.md
@@ -57,9 +57,9 @@ nvm use 6.11.5
 ```diff
 const EventEmitter = require('events').EventEmitter;
 
-+ const dataRetriever = (countTries) => {
++ const dataRetriever = (countries) => {
 +   const emitter = new EventEmitter();
-+   deferredProcess(countTries, emitter);
++   deferredProcess(countries, emitter);
 
 +   return emitter;
 + };
@@ -73,13 +73,13 @@ const EventEmitter = require('events').EventEmitter;
 ```diff
 const EventEmitter = require('events').EventEmitter;
 
-+ const deferredProcess = (countTries, emitter) => {
++ const deferredProcess = (countries, emitter) => {
 +   process.nextTick(() => {
 +     let count = 0;
 +     emitter.emit('start');
 +     const interval = setInterval(() => {
 +       emitter.emit('data', ++count);
-+       if (count === countTries) {
++       if (count === countries) {
 +         emitter.emit('end', count);
 +         clearInterval(interval);
 +       }
@@ -87,9 +87,9 @@ const EventEmitter = require('events').EventEmitter;
 +   });
 + };
 
-const dataRetriever = (countTries) => {
+const dataRetriever = (countries) => {
   const emitter = new EventEmitter();
-  deferredProcess(countTries, emitter);
+  deferredProcess(countries, emitter);
 
   return emitter;
 };
@@ -103,13 +103,13 @@ const dataRetriever = (countTries) => {
 ```diff
 const EventEmitter = require('events').EventEmitter;
 
-const deferredProcess = (countTries, emitter) => {
+const deferredProcess = (countries, emitter) => {
   process.nextTick(() => {
     let count = 0;
     emitter.emit('start');
     const interval = setInterval(() => {
       emitter.emit('data', ++count);
-      if (count === countTries) {
+      if (count === countries) {
         emitter.emit('end', count);
         clearInterval(interval);
       }
@@ -117,9 +117,9 @@ const deferredProcess = (countTries, emitter) => {
   });
 };
 
-const dataRetriever = (countTries) => {
+const dataRetriever = (countries) => {
   const emitter = new EventEmitter();
-  deferredProcess(countTries, emitter);
+  deferredProcess(countries, emitter);
 
   return emitter;
 };
