+++ 03 Events/02 External emitter/README.md
@@ -1,10 +1,14 @@
 # 02 External emitter
 
-In this sample we are going to .
+In this sample we are going to use previous sample, require it from other file and create different instances of `EventEmitter`.
+
 
 Summary steps:
 
--
+- Rename file to `retriever.js`.
+- Create `index.js` calling to `retriever`.
+- Create `EventEmitter` instance using `util.inherits`.
+- Create `EventEmitter` instance `ES6 classes`.
 
 # Steps to build it
 
@@ -14,7 +18,127 @@ Summary steps:
 
 # Steps
 
--
+- Rename file to `retriever.js`.
+
+- Export `retriever` method:
+
+### ./retriever.js
+
+```diff
+const EventEmitter = require('events').EventEmitter;
+
+const deferredProcess = (countries, emitter) => {
+  process.nextTick(() => {
+    let count = 0;
+    emitter.emit('start');
+    const interval = setInterval(() => {
+      emitter.emit('data', ++count);
+      if (count === countries) {
+        emitter.emit('end', count);
+        clearInterval(interval);
+      }
+    }, 300);
+  });
+};
+
+- const dataRetriever = (countries) => {
++ const retriever = (countries) => {
+  const emitter = new EventEmitter();
+  deferredProcess(countries, emitter);
+
+  return emitter;
+};
+
++ module.exports = retriever;
+
+- const dataRetrieverHandler = dataRetriever(10);
+- dataRetrieverHandler.on('start', () => console.log('Started'));
+- dataRetrieverHandler.on('data', (data) => console.log(`Data: ${data}`));
+- dataRetrieverHandler.on('end', (data) => console.log(`End data: ${data}`));
+
+```
+
+- Create `index.js` calling to `retriever`:
+
+### ./index.js
+
+```javascript
+const retriever = require('./retriever');
+
+const retrieverHandler = retriever(10);
+retrieverHandler.on('start', () => console.log('Started'));
+retrieverHandler.on('data', (data) => console.log(`Data: ${data}`));
+retrieverHandler.on('end', (data) => console.log(`End data: ${data}`));
+
+```
+
+- Run app:
+
+```bash
+node index
+```
+
+![extract retriever](../../99%20Resources/03%20Events/02%20External%20emitter/extract%20retriever.gif)
+
+- Other way to create `EventEmitter` instance is extending this class. Previous to `ES6 classes`, Node provides an [`util`](https://nodejs.org/docs/latest-v6.x/api/util.html) module. We are going to use `util.inherits` to extend `EventEmitter`:
+
+### ./retriever.js
+
+```diff
+- const EventEmitter = require('events').EventEmitter;
++ const util = require('util');
+
+- const deferredProcess = (countries, emitter) => {
++ const deferredProcess = (count, countries, emitter) => {
+-   process.nextTick(() => {
+-     let count = 0;
+      emitter.emit('start');
+      const interval = setInterval(() => {
+        emitter.emit('data', ++count);
+        if (count === countries) {
+          emitter.emit('end', count);
+          clearInterval(interval);
+        }
+      }, 300);
+-   });
+};
+
+- const retriever = (countries) => {
++ const Retriever = function(countries) {
+-   const emitter = new EventEmitter();
+-   deferredProcess(countries, emitter);
+
+-   return emitter;
++   const self = this;
++   let count = 0;
++   process.nextTick(() => deferredProcess(count, countries, self));
+};
+
++ util.inherits(
++   Retriever,
++   require('events').EventEmitter
++ );
+
+- module.exports = retriever;
++ module.exports = Retriever;
+
+```
+
+- Update `index` file:
+
+### ./index.js
+
+```diff
+- const retriever = require('./retriever');
++ const Retriever = require('./retriever');
+
+- const retrieverHandler = retriever(10);
++ const retrieverHandler = new Retriever(10);
+retrieverHandler.on('start', () => console.log('Started'));
+retrieverHandler.on('data', (data) => console.log(`Data: ${data}`));
+retrieverHandler.on('end', (data) => console.log(`End data: ${data}`));
+
+```
 
 # About Lemoncode
